+++ ionoscloud/resource_dbaas_inmemorydb_replicaset.go
@@ -104,16 +104,19 @@ func resourceDBaaSInMemoryDBReplicaSet() *schema.Resource {
 							Type:        schema.TypeString,
 							Description: "The datacenter to connect your instance to.",
 							Required:    true,
+							ForceNew:    true,
 						},
 						"lan_id": {
 							Type:        schema.TypeString,
 							Description: "The numeric LAN ID to connect your instance to.",
 							Required:    true,
+							ForceNew:    true,
 						},
 						"cidr": {
 							Type:        schema.TypeString,
 							Description: "The IP and subnet for your instance. Note the following unavailable IP ranges: 10.233.64.0/18, 10.233.0.0/18, 10.233.114.0/24",
 							Required:    true,
+							ForceNew:    true,
 						},
 					},
 				},
@@ -129,12 +132,14 @@ func resourceDBaaSInMemoryDBReplicaSet() *schema.Resource {
 							Type:        schema.TypeString,
 							Description: "The username for the initial InMemoryDB user. Some system usernames are restricted (e.g. 'admin', 'standby').",
 							Required:    true,
+							ForceNew:    true,
 						},
 						"plain_text_password": {
 							Type:         schema.TypeString,
 							Description:  "The password for a InMemoryDB user.",
 							Optional:     true,
 							Sensitive:    true,
+							ForceNew:     true,
 							ExactlyOneOf: []string{"credentials.0.plain_text_password", "credentials.0.hashed_password"},
 						},
 						"hashed_password": {
@@ -148,10 +153,12 @@ func resourceDBaaSInMemoryDBReplicaSet() *schema.Resource {
 									"algorithm": {
 										Type:     schema.TypeString,
 										Required: true,
+										ForceNew: true,
 									},
 									"hash": {
 										Type:     schema.TypeString,
 										Required: true,
+										ForceNew: true,
 									},
 								},
 							},
@@ -184,6 +191,7 @@ func resourceDBaaSInMemoryDBReplicaSet() *schema.Resource {
 				Type:        schema.TypeString,
 				Description: "The ID of a snapshot to restore the replica set from. If set, the replica set will be created from the snapshot.",
 				Optional:    true,
+				ForceNew:    true,
 			},
 			"dns_name": {
 				Type:        schema.TypeString,